switch to ts and addToken
