+++ .gitignore
@@ -11,3 +11,6 @@ contracts/Verifier.sol
 
 contract-deployed.env
 ganache/
+
+# secrets
+/secrets.json
