--- hardhat.config.js
@@ -1,42 +1,0 @@
-const { infuraApiKey, walletPrivateKey, etherscanApiKey } = require('./secrets.json');
-
-require("@nomiclabs/hardhat-etherscan");
-require("@nomiclabs/hardhat-waffle");
-
-// This is a sample Hardhat task. To learn how to create your own go to
-// https://hardhat.org/guides/create-task.html
-task("accounts", "Prints the list of accounts", async () => {
-  const accounts = await ethers.getSigners();
-
-  for (const account of accounts) {
-    console.log(account.address);
-  }
-});
-
-// You need to export an object to set up your config
-// Go to https://hardhat.org/config/ to learn more
-
-/**
- * @type import('hardhat/config').HardhatUserConfig
- */
-module.exports = {
-  defaultNetwork: "hardhat",
-  networks: {
-    hardhat: {
-      chainId: 1337,
-      mining: {
-        auto: false,
-        interval: [3000, 6000]
-      }
-    },
-    goerli: {
-      url: `https://goerli.infura.io/v3/${infuraApiKey}`,
-      accounts: [walletPrivateKey],
-    }
-  },
-  etherscan: {
-    apiKey: etherscanApiKey
-  },
-  solidity: "0.8.4",
-};
-
