+++ package.json
@@ -13,7 +13,9 @@
     "ethers": "^5.0.24",
     "hardhat": "^2.0.6",
     "prettier": "^2.2.1",
-    "prettier-plugin-solidity": "^1.0.0-beta.2"
+    "prettier-plugin-solidity": "^1.0.0-beta.2",
+    "ts-node": "^10.2.1",
+    "typescript": "^4.4.3"
   },
   "dependencies": {
     "@openzeppelin/contracts": "^4.3.2"
