+++ tokens.ts
@@ -1,0 +1,7 @@
+export default [
+    { "name": "USDT", "address": "0x021030D7BA0B21633fE351e83C9b0b8035360329" },
+    { "name": "UNI", "address": "0xB412EcE47485D3CFbf65e1b95408D5679eF9c36d" },
+    { "name": "LINK", "address": "0x1bA66DB7A54738F5597FeCa2640fd2E717722230" },
+    { "name": "YFI", "address": "0xfb30fd66413c01547115C22B62Bde0fF3700DE1F" },
+    { "name": "MATIC", "address": "0x1185f0F4Ea8a713920852b6113F2C3eFe593899F" }
+];
