+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+      "target": "es2018",
+      "module": "commonjs",
+      "strict": true,
+      "esModuleInterop": true,
+      "outDir": "dist",
+      "resolveJsonModule": true
+    },
+    "include": ["./scripts", "./test", "./typechain"],
+    "files": ["./hardhat.config.ts"]
+  }
